+++ ReleaseNotes/release_notes.md
@@ -4,6 +4,9 @@
 
 ## Bug fixes:
 
+  1. fix a bug in `dataset.duplicates` : not all duplicated entries were listed.
+  
+
 # v1.9.9.6
 
 ## New features: